backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1
networkx==3.1
google-generativeai==0.3.0
python-dotenv==1.0.0